Nightly Gaussian Splatting Unit benchmarks (#164)
Contributes to openvdb/fvdb-core#127

Adds a GHA workflow to run unit benchmarks nightly and check for
regressions.
Updates the unit benchmark script to be more robust to differents in
CWD.
Includes utility scripts for downloading and uploading checkpoints
from/to S3.

The upload script can be run manually -- it reads the checkpoints listed
in benchmark_config.yml, uploads them, replaces their paths with S3
URIs, and uploads the config to S3 as well.

The download script is intended to do the opposite, and be used by the
CI workflow. It downloads the config from S3, reads the S3 URIs,
downloads the checkpoints, replaces the URIs with local paths in the
config, and saves it locally.

This config is then used by the test_3dgs.py unit pytests.

---------

Signed-off-by: Mark Harris <mharris@nvidia.com>