Fix nightly benchmark workflow issues (#169)
Fixes the nightly benchmark workflow.

- Gets the commit hash at the beginning of the workflow and uses it
throughput to avoid commits that happen in the middle of the run causing
artifact download to fail. Can also find artifacts from previous runs of
the same hash of fvdb-core, which can help if you manually run the
workflow multiple times.
- Fixes the test_3dgs.py script to rebuild the sfm_scene from the
dataset rather than using it from the checkpoint. This avoids problems
due to the checkpoint containing cache paths that don't exist on the CI
runner.
- Fixes various other errors in the script around permissions, download
paths, etc.
 
---------

Signed-off-by: Mark Harris <mharris@nvidia.com>